### 2021-09-09 #3707 @lootie
License file added to repo (gnu GPL v2.0).
