### 2021-09-27 #3766 @whedon
[REVIEW]: LaplaceInterpolation.jl: A Julia package for fast interpolation on a grid interpolation on a grid 
**Submitting author:** @lootie (<a href="http://orcid.org/0000-0003-3996-773X">Charlotte Haley</a>)
**Repository:** <a href="https://github.com/vishwas1984/LaplaceInterpolation.jl" target ="_blank">https://github.com/vishwas1984/LaplaceInterpolation.jl</a>
**Version:** v0.1.2
