### 2021-10-11 0b6d6b6
Delete difpaper.md
Former-commit-id: 3d48dc3e2eef7b5769033a90d90cba647fb24bd0
Former-commit-id: b0d500a5840978224639b83f6f259314594dbbfb
