+++ paper/difpaper.md
@@ -1,294 +1,0 @@
-1,150c1,142
-< ---
-< title: 'LaplaceInterpolation.jl: A Julia package for fast interpolation on a grid'
-< tags:
-<   - Julia
-<   - statistics
-<   - spatial statistics
-<   - space-time processes
-< authors:
-<   - name: Vishwas Rao
-<     orcid: 0000-0002-4395-6075
-<     affiliation: "1"
-<     email: vhebbur@anl.gov
-<   - name: Charlotte L. Haley
-<     orcid: 0000-0003-3996-773X
-<     affiliation: "1"
-<     email: "haley@anl.gov" 
-<   - name: Mihai Anitescu
-<     affiliation: "1,2"
-< affiliations:
-<  - name: Argonne National Laboratory
-<    index: 1
-<  - name: University of Chicago
-<    index: 2
-< date: 28 June, 2021
-< bibliography: paper.bib
-< 
-< ---
-< 
-< # Summary
-< 
-< We implement a linear-time algorithm for interpolation on a regular
-< multidimensional grid in the Julia language. The algorithm is an
-< %GAIL  - are you  allowed to  use  refs in  the  summary? usually not  in Abstracts
-< approximate Laplace interpolation [@press1992] when no parameters are given; and
-< when parameters $m\in\mathbb{Z}$ and $\epsilon > 0$ are set, the interpolant
-< approximates a Matérn kernel, of which radial basis functions and polyharmonic
-< splines are a special case. We implement, in addition, Neumann, Dirichlet
-< (trivial), and average boundary conditions with potentially different aspect
-< ratios in the different dimensions. The interpolant functions in arbitrary
-< dimensions.
-< 
-< # Mathematical Background
-< 
-< Radial basis functions and splines can be unified conceptually through the
-< notion of Green's functions and eigenfunction expansions [@fasshauer2012green].  The
-< general multivariate Matérn kernels are of the form 
-< 
-< $$ K(\mathbf x ; \mathbf z) = K_{m-d/2}(\epsilon||\mathbf x -\mathbf z ||)(\epsilon||\mathbf x - \mathbf z ||)^{m-d/2}$$
-< 
-< for $m > d/2$, where $K$ is the modified Bessel function of the second kind with parameter $\nu$ and can be
-< obtained as Green’s kernels of 
-< 
-< $$ L = (\epsilon^2I-\Delta)^m , $$
-< 
-< where $\Delta$ denotes the Laplacian operator in $d$ dimensions. Polyharmonic
-< splines, including thin plate splines, are a special case of the above, and this
-< class includes the thin plate splines. 
-< 
-< The discrete gridded interpolation seeks to find an interpolation $u (\mathbf x
-< )$ that satisfies the differential operator in $d$ dimensions on the nodes
-< $\mathbf x_i$ where there is no data and equals $y_i$ everywhere else.
-< Discretely, one solves the matrix problem
-< 
-< $$ \mathbf C  (\mathbf u  - \mathbf y ) - (1 - \mathbf C ) L \mathbf u  = 0 , $$
-< 
-< where $\mathbf{y}$ contains the $y_i$'s and placeholders where there is no data, $L$
-< denotes the discrete matrix operator, and $C$ is a diagonal matrix that indicates 
-< whether node $\mathbf x_i$ is observed. 
-< 
-< In $d-$ dimensions the matrix $A^{(d)}$ of size $M \times M$ expands the 
-< first-order finite difference curvature, and its $(i,j)$th entry is $-1$ when node j is
-< in the set of neighbors of the node $\mathbf x_i$ and has the number of such neighbors on the diagonal. 
[…]
