### 2021-10-25 vishwas1984/LaplaceInterpolation.jl#3 @wkearn
Instantiation of Notebooks fails
Working on my JOSS review over at https://github.com/openjournals/joss-reviews/issues/3766, and I've run into a few problems, so there might be a couple of issues coming your way.

First up: your instructions for instantiating the Notebooks directory fail for me (on master):

```julia
(@v1.6) pkg> activate .
(Notebooks) pkg> instantiate
ERROR: expected package `LaplaceInterpolation [afb1b68e]` to be registered
```

I think that to get around LaplaceInterpolation not being registered, you'll need to commit a Manifest.toml file in the Notebooks directory that points to the LaplaceInterpolation package, which would have an entry like

```
[[LaplaceInterpolation]]
deps = ["LinearAlgebra", "SparseArrays"]
git-tree-sha1 = "79cb4ec931bfd912a0dd4a5e28a5dd6e60568fd4"
repo-rev = "master"
repo-url = ".."
uuid = "afb1b68e-e51f-4e49-a799-0e085170423d"
version = "0.1.0"
```

which points to the parent directory ("..") of Notebooks. There might also be other ways to do this.
