+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#LaplaceInterpolation.jl-Documentation","page":"LaplaceInterpolation.jl Documentation","title":"LaplaceInterpolation.jl Documentation","text":"","category":"section"},{"location":"","page":"LaplaceInterpolation.jl Documentation","title":"LaplaceInterpolation.jl Documentation","text":"This package quickly interpolates data on a grid in one and higher dimensions. ","category":"page"},{"location":"#Installation-Instructions","page":"LaplaceInterpolation.jl Documentation","title":"Installation Instructions","text":"","category":"section"},{"location":"","page":"LaplaceInterpolation.jl Documentation","title":"LaplaceInterpolation.jl Documentation","text":"This package is unregistered, so please install using","category":"page"},{"location":"","page":"LaplaceInterpolation.jl Documentation","title":"LaplaceInterpolation.jl Documentation","text":"Pkg> add https://github.com/vishwas1984/LaplaceInterpolation.jl","category":"page"},{"location":"#Getting-started","page":"LaplaceInterpolation.jl Documentation","title":"Getting started","text":"","category":"section"},{"location":"","page":"LaplaceInterpolation.jl Documentation","title":"LaplaceInterpolation.jl Documentation","text":"Suppose we need to interpolate the following image ","category":"page"},{"location":"","page":"LaplaceInterpolation.jl Documentation","title":"LaplaceInterpolation.jl Documentation","text":"using LaplaceInterpolation, TestImgaes\n\nimg = Float64.(Gray.(testimage(\"mandrill\")))","category":"page"},{"location":"","page":"LaplaceInterpolation.jl Documentation","title":"LaplaceInterpolation.jl Documentation","text":"For illustration purposes, we'll punch a few holes and randomize some data","category":"page"},{"location":"","page":"LaplaceInterpolation.jl Documentation","title":"LaplaceInterpolation.jl Documentation","text":"rows, columns = (256, 512)\nN = rows*columns\n\nmat = convert(Array{Float64}, imgg)[1:rows,1:columns]\n\nN2 = Int64(round(N/2))\nNo_of_nodes_discarded = Int64(round(0.9*N2))\n\ndiscard1 = N2 .+ randperm(N2)[1:No_of_nodes_discarded]\n\ncent = [(150, 150), (60, 100)]\nrad = 30*ones(Int64, 2)\ndiscard2 = punch_holes_2D(cent, rad, rows, columns);\n\ndiscard = vcat(discard1, discard2)\nmat[discard] .= 1\n\nheatmap(mat, title = \"Image with Missing data\", yflip = true, \n              c = :bone, clims = (0.0, 1.0))","category":"page"},{"location":"","page":"LaplaceInterpolation.jl Documentation","title":"LaplaceInterpolation.jl Documentation","text":"Interpolating using the laplace and matern approximations, we get","category":"page"},{"location":"","page":"LaplaceInterpolation.jl Documentation","title":"LaplaceInterpolation.jl Documentation","text":"restored_img_laplace = matern_2d_grid(mat, discard, 1, 0.0)\nrestored_img_matern = matern_2d_grid(mat, discard, 2, 0.0)","category":"page"},{"location":"","page":"LaplaceInterpolation.jl Documentation","title":"LaplaceInterpolation.jl Documentation","text":"And plotting, we have","category":"page"},{"location":"","page":"LaplaceInterpolation.jl Documentation","title":"LaplaceInterpolation.jl Documentation","text":"p1 = heatm
[…]
