+++ README.md
@@ -1,7 +1,7 @@
 # Fast Interpolation for Volume Datasets
 
 [![Build Status](https://github.com/vishwas1984/LaplaceInterpolation.jl/actions/workflows/CI.yml/badge.svg)](https://github.com/vishwas1984/LaplaceInterpolation.jl/actions)
-[![](https://img.shields.io/badge/docs-stable-blue.svg)](https://vishwas1984.github.io/LaplaceInterpolation.jl/stable)
+[![](https://img.shields.io/badge/docs-stable-blue.svg)](https://vishwas1984.github.io/LaplaceInterpolation.jl/dev)
 [![](https://img.shields.io/badge/docs-dev-blue.svg)](https://vishwas1984.github.io/LaplaceInterpolation.jl/dev)
 
 This code performs Laplace and Matern interpolation where missing data are on a one, two, or three
