+++ Notebooks/Project.toml
@@ -26,5 +26,4 @@ SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 TestImages = "5e47fb64-e119-507b-a336-dd2b206d9990"
-ThinPlateSplines = "1d861738-f48e-4029-b1d3-81ce6bc7f5ab"
 VectorizedRoutines = "0e69188a-a5d4-5622-b4e4-a72373136fc5"
