+++ Notebooks/Project.toml
@@ -1,29 +1,10 @@
 [deps]
-AlgebraicMultigrid = "2169fc97-5a83-5252-b627-83903c6c433c"
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
-Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
-Conda = "8f4d0f93-b110-5947-807f-2305c1781a2d"
-Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
-DistributedArrays = "aaf54ef3-cdf8-58ed-94cc-d582ad619b94"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
-GR = "28b8d3ca-fb5f-59d9-8090-bfdbd6d07a71"
 IJulia = "7073ff75-c697-5162-941a-fcdaad2a7d2a"
 JLD = "4138dd39-2aa7-5051-a626-17a0bb65d9c8"
-LaTeXStrings = "b964fa9f-0449-5b57-a5c2-d3ea65f4040f"
 LaplaceInterpolation = "afb1b68e-e51f-4e49-a799-0e085170423d"
-Laplacians = "6f8e5838-0efe-5de0-80a3-5fb4f8dbb1de"
-Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-NPZ = "15e1cf62-19b3-5cfa-8e77-841668bca605"
-NearestNeighbors = "b8a86587-4115-5ab1-83bc-aa920d37bbce"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-Profile = "9abbd945-dff8-562f-b5e8-e1ebf5ef1b79"
-PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
-PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
-QuartzImageIO = "dca85d43-d64c-5e67-8c65-017450d5d020"
-Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
-SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
-SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 TestImages = "5e47fb64-e119-507b-a336-dd2b206d9990"
-VectorizedRoutines = "0e69188a-a5d4-5622-b4e4-a72373136fc5"
