### 2021-11-17 b0e8483
changes to Notebooks Project.toml, Manifest.toml, rerun notebooks
Former-commit-id: 08d3fadd5ec3340caf5a85b7ccb9ebba218733bc
Former-commit-id: 788e5b121ae7857302352fb16933ea888a4e5d4d
