+++ Notebooks/Project.toml
@@ -1,10 +1,18 @@
 [deps]
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
+Conda = "8f4d0f93-b110-5947-807f-2305c1781a2d"
+DSP = "717857b8-e6f2-59f4-9121-6e50c889abd2"
+FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
+GR = "28b8d3ca-fb5f-59d9-8090-bfdbd6d07a71"
+GaussianProcesses = "891a1506-143c-57d2-908e-e1f8e92e6de9"
 IJulia = "7073ff75-c697-5162-941a-fcdaad2a7d2a"
+ImageIO = "82e4d734-157c-48bb-816b-45c225c6df19"
 JLD = "4138dd39-2aa7-5051-a626-17a0bb65d9c8"
 LaplaceInterpolation = "afb1b68e-e51f-4e49-a799-0e085170423d"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+QuartzImageIO = "dca85d43-d64c-5e67-8c65-017450d5d020"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
+StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
 TestImages = "5e47fb64-e119-507b-a336-dd2b206d9990"
