### 2021-11-17 b6691d7
Remove LaplaceInterpolation from the Notebooks
Former-commit-id: 1486a4c18192a2ab50cf6da848e1625545c9c33f
Former-commit-id: 94b18a816f90d87f8ac761f1557a023462eae1e1
