+++ Notebooks/Project.toml
@@ -9,7 +9,6 @@ GaussianProcesses = "891a1506-143c-57d2-908e-e1f8e92e6de9"
 IJulia = "7073ff75-c697-5162-941a-fcdaad2a7d2a"
 ImageIO = "82e4d734-157c-48bb-816b-45c225c6df19"
 JLD = "4138dd39-2aa7-5051-a626-17a0bb65d9c8"
-LaplaceInterpolation = "afb1b68e-e51f-4e49-a799-0e085170423d"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 QuartzImageIO = "dca85d43-d64c-5e67-8c65-017450d5d020"
