+++ README.md
@@ -29,11 +29,11 @@ are located in the `/Notebooks` directory.
 To run the examples yourself, clone the repo, navigate to the Notebooks
 directory, start julia and use
 ```
-julia> using Pkg IJulia 
-julia> Pkg.activate() 
+julia> using Pkg 
+] activate . 
 julia> Pkg.add(url="https://github.com/vishwas1984/LaplaceInterpolation.jl") 
 julia> Pkg.instantiate()
-julia> notebook(dir=".") 
+julia> include("run_notebooks.jl")
 ```
 which will start a jupyter notebook for you, with all relevant dependencies.
 
