### 2022-01-19 #3766 @eviatarbach
Apologies, I did not see that the notebook issue had been resolved. I commend the authors on the great contribution: having a fast interpolation algorithm available will be of interest to many communities. The checklist requirements have to be fully satisfied before it can be published, though. My comments follow.

First of all, the size of the repository is quite big, at 185 MB, due to git blobs (.git/objects/pack alone is 166 MB). Maybe consider using git-filter-branch or something like [BFG](https://rtyley.github.io/bfg-repo-cleaner/).

**Notebooks**

Is there a smaller example you can use for the 3D interpolation in JOSS_Companion_Notebook? When I tried it, it consumed over 5 GB of RAM, which would be too much for many personal computers. At least add a warning to the notebook about this.

**Documentation**

Please add information about the notebooks to the documentation.

In addition, you are missing the following checklist items from the documentation:
> - A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> - Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
> - Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

**Paper**

1. Please clarify the time complexity. Is the algorithm is linear in the number of points to be interpolated? In the volume of the space?
2. Please add a performance section, showing the scaling of the Matérn vs. Laplace interpolation.
2. Please add citations to the other packages you reference in the paper, not just links. Also, the link for gridinterppy is dead.

Minor edits:
L26: "d— dimensions" → "d dimensions"
L27: The j index is not italicized.
L36: "interpolation to solve" → "interpolation to be solved"
L63: "wrapperr" → "wrapper"
