+++ paper/paper.bib
@@ -125,3 +125,145 @@ publisher	= {Cambridge University Press}
   year={2011},
   organization={Springer}
 }
+
+@misc{gridinterppy,
+ author        = {Haley, C. L.},
+ year          = {2021},
+ url           = {https://github.com/lootie/gridinterppy},
+ title         = {{lootie/gridinterppy: Python wrapper to LaplaceInterpolation.jl}}
+}
+
+@misc{Interpolations,
+ author        = {Lycken, T. and Holy T. and contributors},
+ year          = {2014},
+ url           = {https://github.com/JuliaMath/Interpolations.jl},
+ month         = {8},
+ title         = {{JuliaMath/Interpolations.jl: Fast, continuous interpolation of discrete datasets in Julia}}
+}
+
+@misc{GridInterpolations,
+ author        = {Kochenderfer, M. and Stanford Intelligent Systems Laboratory},
+ year          = {2016},
+ url           = {https://github.com/sisl/GridInterpolations.jl},
+ month         = {1},
+ title         = {{sisl/GridInterpolations.jl: Multidimensional grid interpolation in arbitrary dimensions}}
+}
+
+@misc{Dierckx,
+ author        = {Barbary, K. and contributors},
+ year          = {2014},
+ url           = {https://github.com/kbarbary/Dierckx.jl},
+ month         = {July},
+ title         = {{kbarbary/Dierckx.jl: a Julia package for 1-d and 2-d splines}}
+}
+
+@misc{Laplacians,
+ author        = {Spielman, D. and contributors}, 
+ year          = {2015},
+ url           = {https://github.com/danspielman/Laplacians.jl},
+ month         = {6},
+ title         = {{danspielman/Laplacians.jl: Algorithms inspired by graph Laplacians: linear equation solvers, sparsification, clustering, optimization, etc.}}
+}
+
+@article{astropy2013,
+Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+Adsurl = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A},
+Archiveprefix = {arXiv},
+Author = {{Astropy Collaboration} and {Robitaille}, T.~P. and {Tollerud}, E.~J. and {Greenfield}, P. and {Droettboom}, M. and {Bray}, E. and {Aldcroft}, T. and {Davis}, M. and {Ginsburg}, A. and {Price-Whelan}, A.~M. and {Kerzendorf}, W.~E. and {Conley}, A. and {Crighton}, N. and {Barbary}, K. and {Muna}, D. and {Ferguson}, H. and {Grollier}, F. and {Parikh}, M.~M. and {Nair}, P.~H. and {Unther}, H.~M. and {Deil}, C. and {Woillez}, J. and {Conseil}, S. and {Kramer}, R. and {Turner}, J.~E.~H. and {Singer}, L. and {Fox}, R. and {Weaver}, B.~A. and {Zabalza}, V. and {Edwards}, Z.~I. and {Azalee Bostroem}, K. and {Burke}, D.~J. and {Casey}, A.~R. and {Crawford}, S.~M. and {Dencheva}, N. and {Ely}, J. and {Jenness}, T. and {Labrie}, K. and {Lim}, P.~L. and {Pierfederici}, F. and {Pontzen}, A. and {Ptak}, A. and {Refsdal}, B. and {Servillat}, M. and {Streicher}, O.},
+Doi = {10.1051/0004-6361/201322068},
+Eid = {A33},
+Eprint = {1307.6212},
+Journal = {\aap},
[…]
