### 2022-01-19 #3766 @wkearn
My apologies on the delay. Thanks for the updates to the installation instructions and notebooks. Everything seems to work properly now.

This package provides an implementation of a fast algorithm for interpolation on grids of arbitrary dimensions. There are other Julia packages that do various kinds of interpolation, which are cited in this paper, though none as far as I know that implement this particular algorithm.

I don't necessarily think that this should hold up the review, but the examples in the documentation and notebooks of restoring images are somewhat contrived. Maybe there are other applications that would do a better job of illustrating the strength of this interpolation approach that could be included in the documentation?

As @eviatarbach stated, there is no statement of need in the documentation, though you could easily add it. There are, however, automated tests (`Pkg.test("LaplaceInterpolation")` runs them), and a brief set of contributing instructions in the README. It seems like you have already made the changes to the citations and typos.
