### 2022-01-20 #3766 @vishwas1984
@eviatarbach
Thanks for the detailed comments. I have used BFG as suggested to remove the blobs and the size of the repo after blob deletion is 8.27 MB (from this site https://citizendot.github.io/gh-info/). Thanks a lot for the suggestion.

Notebooks:
I have added a toy example for demonstrating 3D case in the notebook and for this problem the memory requirements are much lesser. I have left the previous example as is and I have added a warning about system requirements.
My Colleague @lootie will address the remaining reviews.
