### 2022-01-20 #3766 @lootie
@eviatarbach Thank you for your detailed comments. I have made the changes to the paper that you've suggested, and have recompiled the paper. Additional comments below.

Documentation
> The `docs/src/index.md` file contains the text “The `Notebooks` folder contains this and other examples.“, but has been placed under its own heading, per your suggestion, for visibility.

Checklist items
A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> The statement of need is located in the paper.md itself under its own heading. Thank you.

Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
> The testfile for the software is located in tests/runtests.jl and can be run from the Julia REPL using
`pkg> test LaplaceInterpolation`

> Thank you. The community guidelines are in the README.md file on the github page under “Contributing”, and are as follows: “Contributions are welcome in the form of issues, pull requests. Support questions can be directed to vhebbur at anl.gov.”

The performance comparison I have added to the paper, as per your suggestion, can be reproduced using the notebook, included (Notebooks/Another_1D_Example.ipynb).

@wkearn Thank you for your input on the examples we’ve given. We developed this package in order to interpolate missing data from volumetric images of X-ray scattering. Specifically, an implementation on a 1000x1000x1000 image with identical spherical punches totaling about 1/10 of the dataset was taking half an hour on a local cluster using the astropy.convolve algorithm. Given the number of such datasets, and the desire to process the volumes as a data stream, we wrote this package. This basic use case is shown on the README page, and this algorithm has been written into a much larger software suite for processing X-ray images.  While this example is important to us, we recognize that Laplace interpolation can be used by a much broader set - especially those interested in 2D images, and one finds more and more examples of 3D and 4D gridded datasets (geophysical data and space-time data), that we decided to make the package extensible to arbitrary dimensions.
