### 2022-01-24 #3766 @eviatarbach
Excellent, thank you for the fast reply! The performance comparisons are quite helpful. Everything looks good to me now.
