+++ paper/paper.bib
@@ -15,12 +15,13 @@
   month		= jan
 }
 
-@Book{		  atkinson2005theoretical,
+@Book{atkinson2005theoretical,
   title		= {Theoretical Numerical Analysis},
   publisher	= {Springer},
   year		= {2005},
   author	= {Atkinson, Kendall and Han, Weimin},
-  volume	= {39}
+  volume	= {39},
+  doi={10.1016/c2013-0-12559-7}
 }
 
 @Book{		  cressie1992,
@@ -35,7 +36,8 @@
   author	= {Daubechies, I.},
   volume	= {61},
   year		= {1992},
-  publisher	= {Siam}
+  publisher	= {Siam},
+  doi={}
 }
 
 @Article{	  fasshauer2011reproducing,
@@ -47,17 +49,19 @@
   number	= {3},
   pages		= {585--611},
   year		= {2011},
-  publisher	= {Springer}
+  publisher	= {Springer},
+  doi={10.1007/s00211-011-0391-2}
 }
 
-@InCollection{	  fasshauer2012green,
+@InCollection{fasshauer2012green,
   title		= {Green’s functions: Taking another look at kernel
 		  approximation, radial basis functions, and splines},
   author	= {Fasshauer, Gregory E},
   booktitle	= {Approximation Theory XIII: {S}an {A}ntonio 2010},
   pages		= {37--63},
   year		= {2012},
-  publisher	= {Springer}
+  publisher	= {Springer},
+  doi={10.1007/978-1-4614-0772-0_4}
 }
 
 @InProceedings{	  gardner2018product,
@@ -123,7 +127,8 @@ publisher	= {Cambridge University Press}
   booktitle={International Conference on Scale Space and Variational Methods in Computer Vision},
   pages={26--37},
   year={2011},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-642-24785-9_3}
 }
 
 @misc{gridinterppy,
