+++ paper/paper.md
@@ -173,7 +173,7 @@ The authors would like to thank Ray Osborn, Matt Krogstad, and Stefan Rosenkranz
 for their input on this work. 
 
 This material is based upon work supported by the U.S. Department of Energy,
-Office of Science, Office of Basic Energy Sciences, under contract 
+Office of Science, Office of Basic Energy Sciences, Materials Sciences and Engineering Division, under contract 
 DE-AC02-06CH11357.
 
 The submitted manuscript has been created by UChicago Argonne, LLC, Operator of Argonne National 
