+++ paper/paper.bib
@@ -37,7 +37,7 @@
   volume	= {61},
   year		= {1992},
   publisher	= {Siam},
-  doi={}
+  doi={10.1137/1.9781611970104}
 }
 
 @Article{	  fasshauer2011reproducing,
