### 2022-09-21 #3766 @tarleb
Hi all! We are working towards a more unified archive of all JOSS papers. In this process, I noted that the Markdown sources for this paper contain markup to include an image that does not exist, namely `images/Onedim_GP.png`. It seems that the image was supposed to be a combination of three files in the *figures* folder. For best results, the figure caption should be given as the image description, as in `![figure caption, possibly spanning many lines](figures/Onedim_FP.png)`. @lootie could you take a look?
